Minor adjustments.
- Updated the DaphneDSL language reference.
- Fixed a part of a condition that could never be true (size_t < 0).
- Changed the renaming of the variables (..._Size to ..._Pos) to ..._SizeT in InsertRow.h and InsertCol.h; the suffix _Size may have been misleading, it was meant to express that the variable is of type size_t (as opposed to type VTSel); I hope it's clearer now.
- Fixed the two new script-level test cases for left indexing: we need to insert a matrix (not a scalar) into a matrix; otherwise, running the script fails - but for unintended reasons.

+++ src/runtime/local/kernels/ExtractRow.h
@@ -111,7 +111,7 @@ template <typename VTSel> struct ExtractRow<Frame, Frame, VTSel> {
             VTSel pos = valuesSel[r];
             if (pos < 0)
                 pos += numRowsArg;
-            if (static_cast<size_t>(pos) < 0 || static_cast<size_t>(pos) >= numRowsArg) {
+            if (pos < 0 || static_cast<size_t>(pos) >= numRowsArg) {
                 std::ostringstream errMsg;
                 errMsg << "invalid argument '" << valuesSel[r]
                        << "' passed to ExtractRow: "
